change: move form to jobs/index.js
